requirements.txt
anyio==3.7.1
asgiref==3.7.2
certifi==2023.5.7
click==8.1.6
colorama==0.4.6
dnspython==2.4.0
email-validator==2.0.0.post2
fastapi==0.70.0
h11==0.14.0
httpcore==0.17.3
idna==3.4
pydantic==1.10.11
PyJWT==1.7.1
python-decouple==3.3
sniffio==1.3.0
starlette==0.16.0
typing_extensions==4.7.1
uvicorn==0.15.0
